old_requirements.txt
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
psutil==5.9.4
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
Werkzeug==2.2.2
zipp==3.11.0
